+++ base.py
@@ -1,16 +1,13 @@
-from collections import OrderedDict, abc as container_abcs
-from itertools import chain, islice
-from typing import Iterator
 import matplotlib.pyplot as plt
 import torch
-from torch.nn import Module, Linear
-from tqdm import tqdm
-from torch.optim import Adam
 from torchvision.datasets import MNIST
 from torchvision.transforms import Compose, ToTensor, Normalize, Lambda
 from torch.utils.data import DataLoader
 
-from torch._jit_internal import _copy_to_script_wrapper
+from FFNetwork import FFNetwork
+import FFEncoding
+
+overlay_y_on_x = FFEncoding.overlay
 
 
 def MNIST_loaders(train_batch_size=50000, test_batch_size=10000):
@@ -37,91 +34,6 @@ def MNIST_loaders(train_batch_size=50000, test_batch_size=10000):
     return train_loader, test_loader
 
 
-def overlay_y_on_x(x, y):
-    """Replace the first 10 pixels of data [x] with one-hot-encoded label [y]"""
-    x_ = x.clone()
-    x_[:, :10] *= 0.0
-    x_[range(x.shape[0]), y] = x.max()
-    return x_
-
-
-class Net(torch.nn.Module):
-    def __init__(self, dims):
-        super().__init__()
-        self.layers = []
-        for d in range(len(dims) - 1):
-            self.layers += [Layer(dims[d], dims[d + 1]).cuda()]
-            self.add_module(str(d), Layer(dims[d], dims[d + 1]))
-
-    def __len__(self) -> int:
-        return len(self._modules)
-
-    @_copy_to_script_wrapper
-    def __dir__(self):
-        keys = super().__dir__()
-        keys = [key for key in keys if not key.isdigit()]
-        return keys
-
-    @_copy_to_script_wrapper
-    def __iter__(self) -> Iterator[Module]:
-        return iter(self._modules.values())
-
-    def forward(self, input):
-        for module in self:
-            input = module(input)
-        return input
-
-    def predict(self, x):
-        goodness_per_label = []
-        for label in range(10):
-            h = overlay_y_on_x(x, label)
-            goodness = []
-            for layer in self.layers:
-                h = layer(h)
-                goodness += [h.pow(2).mean(1)]
-            goodness_per_label += [sum(goodness).unsqueeze(1)]
-        goodness_per_label = torch.cat(goodness_per_label, 1)
-        return goodness_per_label.argmax(1)
-
-    def train(self, x_pos, x_neg):
-        h_pos, h_neg = x_pos, x_neg
-        for i, layer in enumerate(self.layers):
-            print("training layer", i, "...")
-            h_pos, h_neg = layer.train(h_pos, h_neg)
-
-
-class Layer(Linear):
-    def __init__(self, in_features, out_features, bias=True, device=None, dtype=None):
-        super().__init__(in_features, out_features, bias, device, dtype)
-        self.relu = torch.nn.ReLU()
-        self.opt = Adam(self.parameters(), lr=0.03)
-        self.threshold = 2.0
-        self.num_epochs = 1000
-
-    def forward(self, x):
-        x_direction = x / (x.norm(2, 1, keepdim=True) + 1e-4)
-        return self.relu(torch.mm(x_direction, self.weight.T) + self.bias.unsqueeze(0))
-
-    def train(self, x_pos, x_neg):
-        for i in tqdm(range(self.num_epochs)):
-            g_pos = self.forward(x_pos).pow(2).mean(1)
-            g_neg = self.forward(x_neg).pow(2).mean(1)
-            # The following loss pushes pos (neg) samples to
-            # values larger (smaller) than the self.threshold.
-            loss = torch.log(
-                1
-                + torch.exp(
-                    torch.cat([-g_pos + self.threshold, g_neg - self.threshold])
-                )
-            ).mean()
-            self.opt.zero_grad()
-            # this backward just compute the derivative and hence
-            # is not considered backpropagation.
-            loss.backward()
-            self.opt.step()
-        return self.forward(x_pos).detach(), self.forward(x_neg).detach()
-
-
 def visualize_sample(data, name="", idx=0):
     reshaped = data[idx].cpu().reshape(28, 28)
     plt.figure(figsize=(4, 4))
@@ -130,25 +42,33 @@ def visualize_sample(data, name="", idx=0):
     plt.show()
 
 
-if __name__ == "__main__":
-    torch.manual_seed(1234)
-    train_loader, test_loader = MNIST_loaders()
+def training_loop(model, x, y, encoding="overlay"):
+    x_pos, x_neg = None, None
+    if encoding == "overlay":
+        x_pos = overlay_y_on_x(x, y)
+        rnd = torch.randperm(x.size(0))
+        x_neg = overlay_y_on_x(x, y[rnd])
 
-    net = Net([784, 500, 500])
-    x, y = next(iter(train_loader))
-    x, y = x.cuda(), y.cuda()
-    x_pos = overlay_y_on_x(x, y)
-    rnd = torch.randperm(x.size(0))
-    x_neg = overlay_y_on_x(x, y[rnd])
+    h_pos, h_neg = x_pos, x_neg
+    # output = model(x_pos, x_neg, encoding)
+    for i, module in enumerate(model.children()):
+        print("Training layer", i, "...")
+        h_pos, h_neg = module.train(h_pos, h_neg)
 
-    for data, name in zip([x, x_pos, x_neg], ["orig", "pos", "neg"]):
-        visualize_sample(data, name)
 
-    net.train(x_pos, x_neg)
+torch.manual_seed(1234)
+train_loader, test_loader = MNIST_loaders()
 
-    print("train error:", 1.0 - net.predict(x).eq(y).float().mean().item())
+net = FFNetwork([784, 500, 500])
+net = net.cuda()
+x, y = next(iter(train_loader))
+x, y = x.cuda(), y.cuda()
 
-    x_te, y_te = next(iter(test_loader))
-    x_te, y_te = x_te.cuda(), y_te.cuda()
+training_loop(net, x, y)
 
-    print("test error:", 1.0 - net.predict(x_te).eq(y_te).float().mean().item())
+print("train error:", 1.0 - net.predict(x).eq(y).float().mean().item())
+
+x_te, y_te = next(iter(test_loader))
+x_te, y_te = x_te.cuda(), y_te.cuda()
+
+print("test error:", 1.0 - net.predict(x_te).eq(y_te).float().mean().item())
